requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
sqlalchemy>=2.0,<3.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0
openai>=1.0
psycopg[binary]>=3.0,<4.0
python-docx>=1.0
google-api-python-client>=2.0
google-auth>=2.0
google-auth-oauthlib>=1.0
gspread>=6.0
